requirements.txt
line-bot-sdk
bs4
flask
pymongo
datetime
pandas
pandas_datareader
SnowNLP
emoji
pyshorteners
scipy
mplfinance
pyimgur
regex
matplotlib
numpy